Add mean pooling example to reducer README
