+++ lightstream/core/reducer/__init__.py
@@ -1,0 +1,10 @@
+from .base import StreamingReducer
+from .gem import StreamingGeMReducer
+from .mean import Reducer, StreamingMeanReducer
+
+__all__ = [
+    "Reducer",
+    "StreamingReducer",
+    "StreamingMeanReducer",
+    "StreamingGeMReducer",
+]
